EnKF.MultiSite

Given the Forecast and Observed this function performs the Ensemble Kalamn Filter.

Usage

EnKF.MultiSite(settings, Forecast, Observed, H, extraArg = NULL, ...)

Arguments

settings

pecan standard settings list.

Forecast

A list containing the forecasts variables including Q (process variance) and X (a dataframe of forecasts state variables for different ensemble)

Observed

A list containing the observed variables including R (cov of observed state variables) and Y (vector of estimated mean of observed state variables)

H

is a matrix of 1's and 0's specifying which observations go with which state variables.

extraArg

This argument is NOT used inside this function but it is a list containing aqq, bqq and t. The aqq and bqq are shape parameters estimated over time for the proccess covariance and t gives the time in terms of index of obs.list.

...

Extra argument sent to the analysis function.

Value

It returns a list with estimated mean and cov matrix of forecast state variables as well as mean and cov estimated as a result of assimilation/analysis .

Details

This function is different than `EnKF` function in terms of how it creates the Pf matrix.

Author

Michael Dietze dietze@bu.edu, Ann Raiho and Hamze Dokoohaki