SDA Downscale Function for Hourly Data

This function uses the randomForest model to downscale forecast data (hourly) to unmodeled locations using covariates and site locations

Usage

SDA_downscale_hrly(nc_file, coords, yyyy, covariates)

Arguments

nc_file

In quotes, file path for .nc containing ensemble data.

coords

In quotes, file path for .csv file containing the site coordinates, columns named "lon" and "lat".

yyyy

In string, format is yyyy(year of interest)

covariates

SpatRaster stack, used as predictors in randomForest. Layers within stack should be named. Recommended that this stack be generated using 'covariates' instructions in assim.sequential/inst folder

Value

It returns the `downscale_output` list containing lists for the training and testing data sets, models, and predicted maps for each ensemble member.

Author

Harunobu Ishii