qsub_sda

This function provides means to split large SDA runs into separate `qsub` jobs. Including job creation, submission, and assemble.

Usage

qsub_sda(
  settings,
  obs.mean,
  obs.cov,
  Q,
  pre_enkf_params,
  ensemble.samples,
  outdir,
  control,
  block.index = NULL
)

Arguments

settings

PEcAn settings object

obs.mean

Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation means for each state variables of each site for each time point.

obs.cov

Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation covariances for all state variables of each site for each time point.

Q

Process covariance matrix given if there is no data to estimate it.

pre_enkf_params

Used for passing pre-existing time-series of process error into the current SDA runs to ignore the impact by the differences between process errors.

ensemble.samples

Pass ensemble.samples from outside to avoid GitHub check issues.

outdir

Physical path to the folder where the SDA outputs will be stored. The default is NULL, where we will be using outdir from the settings object.

control

List of flags controlling the behavior of the SDA. `TimeseriesPlot` for post analysis examination; `OutlierDetection` decide if we want to execute the outlier detection each time after the model forecasting; `send_email` contains lists for sending email to report the SDA progress; `keepNC` decide if we want to keep the NetCDF files inside the out directory; `forceRun` decide if we want to proceed the Bayesian MCMC sampling without observations; `MCMC.args` include lists for controling the MCMC sampling process (iteration, nchains, burnin, and nthin.).

block.index

list of site ids for each block, default is NULL. This is used when the localization turns on. Please keep using the default value because the localization feature is still in development.

Value

NONE

Author

Dongchen Zhang