sda.enkf_local

This function provides complete support for the multi-core and multi-node computation on the general HPC system. Thus, this script will be more computationally efficient, making it possible to run SDA over thousands of locations.

Usage

sda.enkf_local(
  settings,
  obs.mean,
  obs.cov,
  Q = NULL,
  pre_enkf_params = NULL,
  ensemble.samples = NULL,
  outdir = NULL,
  control = list(TimeseriesPlot = FALSE, OutlierDetection = FALSE, send_email = NULL,
    keepNC = TRUE, forceRun = TRUE, MCMC.args = NULL, merge_nc = TRUE),
  debias = list(cov.dir = NULL, start.year = NULL)
)

Arguments

settings

PEcAn settings object

obs.mean

Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation means for each state variables of each site for each time point.

obs.cov

Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation covariances for all state variables of each site for each time point.

Q

Process covariance matrix given if there is no data to estimate it.

pre_enkf_params

Used for passing pre-existing time-series of process error into the current SDA runs to ignore the impact by the differences between process errors.

ensemble.samples

list of ensemble parameters across PFTs. Default is NULL.

outdir

physical path to the folder that stores the SDA outputs. Default is NULL.

control

List of flags controlling the behavior of the SDA. `TimeseriesPlot` for post analysis examination; `OutlierDetection` decide if we want to execute the outlier detection each time after the model forecasting; `send_email` contains lists for sending email to report the SDA progress; `keepNC` decide if we want to keep the NetCDF files inside the out directory; `forceRun` decide if we want to proceed the Bayesian MCMC sampling without observations; `MCMC.args` include lists for controling the MCMC sampling process (iteration, nchains, burnin, and nthin.). `merge_nc` determine if we want to merge all netCDF files across sites and ensembles. If it's set as `TRUE`, we will then combine all netCDF files into the `merged_nc` folder within the `outdir`.

debias

List: R list containing the covariance directory and the start year. covariance directory should include GeoTIFF files named by year. start year is numeric input which decide when to start the debiasing feature.

Value

NONE

Author

Dongchen Zhang zhangdc@bu.edu