sda_assemble

This function can help to assemble sda outputs (analysis and forecasts) from each job execution.

Usage

sda.qsub.job.submission(
  batch.folder,
  username,
  outdir = NULL,
  past.job.ids = NULL,
  max.job = 20,
  prefix = "sda.all.forecast.analysis.Rdata",
  resources = list(hour = 24, memory = "4G", ncpu = 28)
)

Arguments

batch.folder

character: path where the SDA batch jobs stored.

username

character: username for the HPC.

outdir

character: path where to stored the assembled results. Default is NULL.

past.job.ids

vector: vector of the past submitted job ids. Default is NULL.

max.job

numeric: maximum allowance for the number of running jobs on HPC.

prefix

character: file name to be determined if a job is finished or not.

resources

list: computation resources (time, memory, cores) when used to submit jobs.

Value

vector or numeric: return -1 when all jobs are finished; return vector of submitted job ids when there is any job that is not finished.

Author

Dongchen Zhang.