tobit_model_censored

tobit_model_censored

Usage

tobit_model_censored(settings, X, var.names, mu.f, Pf, t)

Arguments

settings

(list) pecan standard settings list.

X

(numeric) A matrix contains ensemble forecasts (ensembles * variables).

var.names

(character) variable names.

mu.f

(numeric) forecast mean values.

Pf

(numeric) forecast covariance matrix.

t

(numeric) timestep. If t=1, initial values are imputed for zero values in mu.f

Value

list with updated mu.f, pf, X, and indication of which y values are censored