SDA_control

SDA_control

Usage

SDA_control(
  trace = TRUE,
  ForewardForecast = FALSE,
  interactivePlot = FALSE,
  TimeseriesPlot = FALSE,
  BiasPlot = FALSE,
  plot.title = NULL,
  facet.plots = FALSE,
  debug = FALSE,
  pause = FALSE,
  Profiling = FALSE,
  OutlierDetection = FALSE
)

Arguments

trace

Logical if code should print out the progress of SDA .

ForewardForecast

Logical if the foreward forecast estimates needs to be read and visualized in time series plots.

interactivePlot

Logical if time series plots need to be generated.

TimeseriesPlot

Logical if time series plots need to be generated.

BiasPlot

Logical if bias plots need to be generated

plot.title

Character defining the title of times series plots

facet.plots

Logical if the timeseries plots should be faceted based on state variables

debug

Logical if the code should stop at some milestones and open an interactive debugging environment

pause

Logical if code needs to be paused and wait for further instruction after the analysis step

Profiling

Logical if code should keep track of how much time each step took

OutlierDetection

Logical if TRUE then a simple method will be used to replace simulations of outside 3IQR with the median of ensembles.

Value

list of all arguments needed to setup the SDA function