Internal functions for plotting SDA outputs. Interactive, post analysis time-series and bias plots in base plotting system and ggplot

Internal functions for plotting SDA outputs. Interactive, post analysis time-series and bias plots in base plotting system and ggplot

Usage

interactive.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.timeser.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.bias.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.bias.plotting.sda.corr(t, obs.times, X, aqq, bqq)

post.analysis.ggplot(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS,
  plot.title = NULL
)

post.analysis.ggplot.violin(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS,
  plot.title = NULL
)

post.analysis.multisite.ggplot(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  FORECAST,
  ANALYSIS,
  plot.title = NULL,
  facetg = FALSE,
  readsFF = NULL,
  Add_Map = FALSE
)

SDA_timeseries_plot(
  ANALYSIS,
  FORECAST,
  obs.mean = NULL,
  obs.cov = NULL,
  outdir,
  pft.path = NULL,
  by = "site",
  types = c("FORECAST", "ANALYSIS", "OBS"),
  CI = c(0.025, 0.975),
  unit = list(AbvGrndWood = "Mg/ha", LAI = "m2/m2", SoilMoistFrac = "", TotSoilCarb =
    "kg/m2"),
  style = list(general_color = c(FORECAST = "blue", ANALYSIS = "red", OBS = "black"),
    fill_color = c(FORECAST = "yellow", ANALYSIS = "green", OBS = "grey"), title_color =
    "red"),
  PDF_w = 20,
  PDF_h = 16,
  t.inds = NULL
)

Arguments

settings

pecan standard settings list.

t

current time - int number giving the position of the current time in obs.time.

obs.times

vector of dates of measurements

obs.mean

obs.mean

obs.cov

obs.cov

obs

list containing the mean and cov object

X

dataframe of state variables for each ensemble

FORECAST

Forecast object from the sda.output.Rdata.

ANALYSIS

Analysis object from the sda.output.Rdata.

aqq, bqq

shape parameters estimated over time for the process covariance

plot.title

character giving the title for post visualization ggplots

facetg

logical: Create a subpanel for each variable?

readsFF

optional forward forecast

Add_Map

Bool variable decide if we want to export the GIS map of Ecoregion.

outdir

physical path where the pdf will be stored.

pft.path

Physical path of pft.csv file to allow by = pft option.

by

arrange figures by var, pft, or site.

types

data types that shown in the figure.

CI

range of confidence interval.

unit

list of unit used for y axis label.

style

color option.

PDF_w

width of exported PDF file, passed on to `base::pdf()`.

PDF_h

height of exported PDF file, passed on to `base::pdf()`.

t.inds

index of period that will be plotted.

Author

Dongchen Zhang